+++ ruleengine-core/src/main/kotlin/ruleengine/evaluator/context/MapRuleContext.kt
@@ -31,16 +31,16 @@ class MapRuleContext(private val map: Map<String, Any?>) : RuleContext {
             is Collection<*> -> {
                 val results = current.mapNotNull { element ->
                     val resolved = resolveRaw(current = element, path = path, index = index)
-                    if (resolved == null) null else resolved
+                    resolved
                 }
-                if (results.isEmpty()) null else results
+                results.ifEmpty { null }
             }
             is Array<*> -> {
                 val results = current.mapNotNull { element ->
                     val resolved = resolveRaw(current = element, path = path, index = index)
-                    if (resolved == null) null else resolved
+                    resolved
                 }
-                if (results.isEmpty()) null else results
+                results.ifEmpty { null }
             }
             else -> null
         }